backend/requirements.txt
alembic==1.17.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
cloudinary==1.44.1
contourpy==1.3.3
cycler==0.12.1
dnspython==2.8.0
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
markup==0.2.2
MarkupSafe==3.0.3
matplotlib==3.10.7
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.2.7
MySQLX==2.3.0
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
pymongo==4.15.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.3
python-socketio==5.14.2
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.44
SQLExecutorX==1.8.6
SQLExecX==2.2.0
typing_extensions==4.15.0
urllib3==2.5.0
uWSGI==2.0.31
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.2.0
